MyDesk/requirements.txt
aiortc==1.14.0
av==16.1.0
firebase_admin==6.4.0
Flask==3.1.2
Flask_Cors==4.0.1
keyring==25.6.0
keyrings.alt==5.0.2
msgpack==1.1.2
mss==10.1.0
numpy==2.4.2
opencv_python_headless==4.13.0.92
Pillow==12.1.0
psutil==7.1.3
PyAudio==0.2.14
pyaudiowpatch==0.2.12.8
pyperclip==1.11.0
PyQt6==6.10.2
pyqt6_sip==13.10.2
python-dotenv==1.2.1
pywin32==311
requests==2.32.5
websockets==15.0.1
winrt==1.0.21033.1
